Implementado el funcionamiento mediante script

+++ src/main/java/com/inigoserrano/bigweather/util/ReducerWeather.java
@@ -9,14 +9,15 @@ public class ReducerWeather implements Reducer<String, String> {
 	@Override
 	public String reduce(final String key, final Iterator<String> iter) {
 
-		int maximo = 0;
+		float maximo = 0;
 		while (iter.hasNext()) {
 			String string = iter.next();
-			if (Integer.parseInt(string) > maximo) {
-				maximo = Integer.parseInt(string);
+			// string = string.replace('.', ',');
+			if (Float.parseFloat(string) > maximo) {
+				maximo = Float.parseFloat(string);
 			}
 		}
-		return Integer.toString(maximo);
+		return Float.toString(maximo);
 	}
 
 }
